tweaking more of the physics engine but problem has arisen
The mesh is not displaying. The world calls and the camera matrix calls are correct. But for some reason, it does not want to work. Might need to check how information is being parsed from the game object code to the constant buffer

+++ Physics Framework/DX11PhysicsFramework/Quaternion.h
@@ -164,12 +164,12 @@ struct Quaternion4
 
 	static float DegreesToRadians(float deg)
 	{
-		return deg *  PI / 180.0f;
+		return float(deg *  PI / 180.0f);
 	}
 
 	static float RadiansToDegrees(float rad)
 	{
-		return rad * 180.0f / PI;
+		return float(rad * 180.0f / PI);
 	}
 
 	static Quaternion4 MakeQFromEulerAngles(float x, float y, float z)
